Use Global AuthenticationManagerBuilder
Using protected configure(AuthenticationManagerBuilder) is typically used for
configuring authentication that is scoped only to this HttpSecurity instance
(in XML HttpSecurity is <http>). Think of protected as being like an inner bean.

However, there is an approach using @Autowired on the AuthenticationManagerBuilder
that allows configuring "global authentication". By "global" it means that
method security can detect the authentication automatically. It also  means that
you can also @Autowired an AuthenticationManager with no additional work. In short,
by using the Global approach you do not need to override authenticationManagerBean.