Enable CSRF
We should not disable CSRF protection. This commit enables CSRF.

Some commentary on the configuration as it was.

1) In order to get the CSRF automatically included we need to use @EnableWebMvcSecurity (notice Mvc). This
provides the RequestDataValueProcessor used by Thymeleaf and the Spring taglibs to automatically include
the hidden input with the CSRF token.

2) The configuraton was:

http
    ...
    .csrf()
        .requireCsrfProtectionMatcher(new AntPathRequestMatcher("/oauth/authorize")).disable()

The configuration is probably better read as:

http
   ...
   .csrf()
     .requireCsrfProtectionMatcher(new AntPathRequestMatcher("/oauth/authorize"))
     .disable()

So what is it actually doing?

a) The requireCsrfProtectionMatcher statement says CSRF will only be applied for the RequestMatcher passed in.
b) The disable statement says that CSRF will be disabled

This means that the first statement actually did nothing since we disable all of CSRF afterwards. This shows
the importance of spacing so that we read the configuration. Note as the indentation is in the example above
it demonstrates that both requireCsrfProtectionMatcher and disable impact csrf. The way expressed previously
it appeared that disable impacted the requireCsrfProtectionMatcher.

How do you know which is which? The value returned from requireCsrfProtectionMatcher is the same type as what
is returned by csrf.

In hindsight is this a bit confusing? Yes...are we stuck with it...yes. I have seen simpilar problems with
others and the only comfort upon the decision of the DSL is that my DSL has the same problems as any language.
The reality is spacing matters in any language, including this DSL. For exmaple I could write

at tributes

or

attributes

They have different meanings based on the spacing.