Remove loginProcessingUrl
The loginPage configures the page that should render the log in form when a GET is
requested. So in this instance GET /login should render the log in page. The default
loginProcessingUrl is a POST to the loginPage. So in this instance POST /login is
the default loginProcessingUrl. This means this configuration is not necessary.