Add anyRequest().authenticated() mapping
The configuration was:

http
   .authorizeRequests()
       .antMatchers("/login").permitAll()
       .and()
   ...

This means there were actually no URL based restrictions in place. Specifically the
following would happen:

Security would say does the URL match /login ?

  - If so, it would grant access and stop.
  - If not, it would continue and see there are no additional mappings and grant access.

We changed the configuration to be:

http
   .authorizeRequests()
       .antMatchers("/login").permitAll()
       .anyRequest().authenticated()
       .and()
   ...

This means that if the mapping does not match /login it requires authentication.

One important thing to note is that if the access denied page was not
"/login?auth_error=true", we could do the following instead:

http
   .authorizedRequests()
      .antMatchers("/denied").permitAll()
      .anyRequest().authenticated()
   .formLogin()
      .loginPage("/login")
      .permitAll()

This grants access to every URL associated with login. However,
since our access denied page is the same as our log in page (with
a different query param) the formLogin().permitAll() is not necessary.
This is because antMatchers does not consider query parameters, but
formLogin().permitAll() does.